feat: Implement condition to allow only todo creator perform read, update and delete on a todo

+++ app/Http/Controllers/TodoController.php
@@ -6,7 +6,6 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth; 
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
-use App\Http\Controllers\Controller;
 use App\Models\Todo;
 use App\Http\Resources\Todo as TodoResource;
 
@@ -19,8 +18,10 @@ class TodoController extends Controller
      */
     public function index()
     {
-        $todos = Todo::paginate(10);
-        return TodoResource::collection($todos);
+        $todos = Todo::where('user_id', auth()->user()->id)->paginate(10);
+        return (TodoResource::collection($todos))
+        ->response()
+        ->setStatusCode(200);;
     }
 
     /**
@@ -43,8 +44,7 @@ class TodoController extends Controller
             ], 422);
         } 
 
-        $todo = $request->isMethod('put') ? 
-        Todo::findOrFail($request->todo_id) : new Todo; 
+        $todo = new Todo; 
 
         $todo->id               =   $request->input('todo_id');
         $todo->user_id          =   $request->user()->id; //get user id from request  
@@ -54,8 +54,72 @@ class TodoController extends Controller
         $todo->slug             =   Str::slug($request->input('title'), '_');
 
         if($todo->save()){
-            return new TodoResource($todo);
+            return (new TodoResource($todo))
+            ->response()
+            ->setStatusCode(201);
         }
+    }
+
+     /**
+     * Update a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'todo_id' => 'required|int',
+        ]);
+        if ($validator->fails())
+        {
+            return response([
+                'errors'=>$validator->errors()->all(),
+                'status' => false,
+                'data' => null
+            ], 422);
+        } 
+
+        $todo = Todo::where('user_id', auth()->user()->id)
+                ->where('id', $request->input('todo_id'))->first();
+        
+        if($todo){
+            
+            $todo->id                   =   $request->input('todo_id');
+            $todo->user_id              =   $request->user()->id; //get user id from request 
+
+            if ($request->has('title')) {
+                $todo->title            =   $request->input('title');
+                $todo->slug             =   Str::slug($request->input('title'), '_');
+            }
+            if ($request->has('description')) {
+                $todo->description      =   $request->input('description');
+            }
+            if ($request->has('is_completed')) {
+                $validator = Validator::make($request->all(), [
+                    'is_completed' => 'boolean',
+                ]);
+                if ($validator->fails())
+                {
+                    return response([
+                        'errors'=>$validator->errors()->all(),
+                        'status' => false,
+                        'data' => null
+                    ], 422);
+                }
+                $todo->is_completed            =   $request->input('is_completed');
+            }  
+    
+            if($todo->save()){
+                return (new TodoResource($todo))
+                ->response()
+                ->setStatusCode(200);
+            }
+        }
+        
+        return (new TodoResource($todo))
+        ->response()
+        ->setStatusCode(403);
     }
 
     /**
@@ -66,8 +130,16 @@ class TodoController extends Controller
      */
     public function show($id)
     {
-        $todo = Todo::findOrFail($id);
-        return new TodoResource($todo);
+        $todo = Todo::where(['user_id' => auth()->user()->id, 'id' => $id])->first();
+        if($todo){
+            return (new TodoResource($todo))
+            ->response()
+            ->setStatusCode(200);
+        }
+
+        return (new TodoResource($todo))
+        ->response()
+        ->setStatusCode(403);        
     }
 
     /**
@@ -78,11 +150,19 @@ class TodoController extends Controller
      */
     public function destroy($id)
     {
-        $todo = Todo::findOrFail($id);
+        $todo = Todo::where(['user_id' => auth()->user()->id, 'id' => $id])->first();
 
-        if($todo->delete()){
-            return new TodoResource($todo);
+        if($todo){
+            if($todo->delete()){
+                return (new TodoResource($todo))
+                ->response()
+                ->setStatusCode(200);
+            }
         }
+
+        return (new TodoResource($todo))
+        ->response()
+        ->setStatusCode(403); 
     }
 
 }
